requirements.txt
asgiref==3.8.1
Django==5.0.6
django-browser-reload==1.13.0
django-ckeditor==6.7.1
django-jazzmin==3.0.0
django-js-asset==2.2.0
gunicorn==22.0.0
mssql-django==1.5
packaging==24.1
pillow==10.4.0
pyodbc==5.1.0
python-slugify==8.0.4
pytz==2024.1
sqlparse==0.5.0
text-unidecode==1.3
whitenoise==6.7.0
